Work around Safari animation bug on completed boards
In Safari, especially on iPhone, scrolling around the page looks flickery and glitchy, when you have completed boards visible.  Worse, when you complete a board, the animation causes the page to stall for several seconds -- and then you don't even see the animation.

This appears to be triggered by the `filter` property on completed boards.

Workaround: if we force 3D acceleration by adding `translateZ(0)` to each board, all is well. There are no glitches, the animation looks good, and the app doesn't stall. It's also more responsive when entering letters.

(There is no need to do this when "hide completed boards" is on, but "animate hiding" is off, so we don't.)

Fixes #24.